# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Encois Gateway API
  version: 1.0.0
paths:
...
  /api/v1/investigations:
    get:
      operationId: listSavedInvestigations
      summary: List saved investigations owned by the caller
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Saved investigation projections
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedInvestigationListResponse'
    post:
      operationId: createSavedInvestigation
      summary: Save a bounded graph, memory, or workflow investigation query
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavedInvestigationCreateRequest'
      responses:
        '201':
          description: Saved investigation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedInvestigationResponse'
  /api/v1/investigations/{investigationId}:
    delete:
      operationId: deleteSavedInvestigation
      summary: Delete a saved investigation owned by the caller
      security:
        - bearerAuth: []
      parameters:
        - name: investigationId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Saved investigation deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletedResponse'
        '404':
          description: Saved investigation not found
  /api/v1/notifications:
    get:
      operationId: listNotifications
      summary: List current user's in-product operational notifications
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Notification projections
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationListResponse'
  /api/v1/notifications/{notificationId}/read:
    post:
      operationId: markNotificationRead
      summary: Mark a notification read
      security:
        - bearerAuth: []
      parameters:
        - name: notificationId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
  /api/v1/settings/notifications:
    get:
      operationId: getNotificationPreferences
      summary: Get current user's notification preferences
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Notification preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreferencesResponse'
    put:
      operationId: updateNotificationPreferences
      summary: Update current user's notification preferences
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationPreferences'
      responses:
        '200':
          description: Notification preferences updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreferencesResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    serviceTokenAuth:
      type: apiKey
      in: header
      name: X-Encois-Service-Token
...
  schemas:
...
    DeletedResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [deleted]
          properties:
            deleted: { const: true }
    AcceptedResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [read]
          properties:
            read: { const: true }
...
    SavedInvestigation:
      type: object
      required: [id, organizationId, name, kind, query, params, scope, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        organizationId: { type: string }
        name: { type: string }
        kind: { type: string, enum: [graph, memory, workflow] }
        query: { type: string }
        params: { type: object, additionalProperties: true }
        scope: { $ref: '#/components/schemas/KnowledgeSourceScope' }
        createdAt: { type: string }
        updatedAt: { type: string }
    SavedInvestigationCreateRequest:
      type: object
      additionalProperties: false
      required: [name, kind, query, scope]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        kind: { type: string, enum: [graph, memory, workflow] }
        query: { type: string, minLength: 1, maxLength: 500 }
        params: { type: object, additionalProperties: true }
        scope: { $ref: '#/components/schemas/KnowledgeSourceScope' }
    SavedInvestigationResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/SavedInvestigation' }
    SavedInvestigationListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/SavedInvestigation' }
    Notification:
      type: object
      required: [id, type, severity, title, message, createdAt]
      properties:
        id: { type: string, format: uuid }
        type: { type: string }
        severity: { type: string, enum: [info, warning, error] }
        title: { type: string }
        message: { type: string }
        resourceType: { type: string }
        resourceId: { type: string }
        readAt: { type: string }
        createdAt: { type: string }
    NotificationListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Notification' }
    NotificationPreferences:
      type: object
      required: [emailEnabled, pushEnabled, workflowUpdates, evidenceReady, weeklyDigest]
      properties:
        emailEnabled: { type: boolean }
        pushEnabled: { type: boolean }
        workflowUpdates: { type: boolean }
        evidenceReady: { type: boolean }
        weeklyDigest: { type: boolean }
        updatedAt: { type: string, readOnly: true }
    NotificationPreferencesResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/NotificationPreferences' }
...
    KnowledgeSourceScope:
      type: object
      additionalProperties: false
      required: [ids]
      properties:
        ids:
          type: array
          minItems: 1
          items: { type: string, minLength: 1 }
...